apps/api/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
asyncpg==0.31.0
python-dotenv==1.1.0
httpx==0.28.1
pydantic==2.6.1
python-multipart==0.0.9
anthropic>=0.30.0
PyJWT==2.8.0
